+++ soup/RenderTarget.cpp
@@ -15,16 +15,16 @@ namespace soup
 		drawRect(0, 0, width, height, colour);
 	}
 
-	void RenderTarget::drawPixel(size_t x, size_t y, Rgb colour)
+	void RenderTarget::drawPixel(unsigned int x, unsigned int y, Rgb colour)
 	{
 		drawRect(x, y, 1, 1, colour);
 	}
 
-	void RenderTarget::drawHollowCircle(size_t x, size_t y, float r, Rgb colour)
+	void RenderTarget::drawHollowCircle(unsigned int x, unsigned int y, float r, Rgb colour)
 	{
 		for (float i = 0.01f; i < M_TAU; i += 0.01f)
 		{
-			drawRect(x + (size_t)(cos(i) * r), y + (size_t)(sin(i) * r), 1, 1, colour);
+			drawRect(x + (unsigned int)(cos(i) * r), y + (unsigned int)(sin(i) * r), 1, 1, colour);
 		}
 	}
 
@@ -44,7 +44,7 @@ namespace soup
 		}
 	}
 
-	static void processPoint(std::map<size_t, std::pair<size_t, size_t>>& lines, size_t x, size_t y)
+	static void processPoint(std::map<unsigned int, std::pair<unsigned int, unsigned int>>& lines, unsigned int x, unsigned int y)
 	{
 		if (auto e = lines.find(y); e != lines.end())
 		{
@@ -59,11 +59,11 @@ namespace soup
 		}
 		else
 		{
-			lines.emplace(y, std::pair<size_t, size_t>{ x, x });
+			lines.emplace(y, std::pair<unsigned int, unsigned int>{ x, x });
 		}
 	}
 
-	static void processLine(std::map<size_t, std::pair<size_t, size_t>>& lines, Vector2 a, Vector2 b)
+	static void processLine(std::map<unsigned int, std::pair<unsigned int, unsigned int>>& lines, Vector2 a, Vector2 b)
 	{
 		for (float t = 0.0f; t < 1.0f; t += (0.5f / a.distance(b)))
 		{
@@ -73,7 +73,7 @@ namespace soup
 
 	void RenderTarget::drawTriangle(Vector2 a, Vector2 b, Vector2 c, Rgb colour)
 	{
-		std::map<size_t, std::pair<size_t, size_t>> lines{};
+		std::map<unsigned int, std::pair<unsigned int, unsigned int>> lines{};
 		processLine(lines, a, b);
 		processLine(lines, b, c);
 		processLine(lines, c, a);
@@ -90,21 +90,21 @@ namespace soup
 		drawLine(c, a, colour);
 	}
 
-	void RenderTarget::drawText(size_t x, size_t y, const std::string& text, const RasterFont& font, Rgb colour, uint8_t scale)
+	void RenderTarget::drawText(unsigned int x, unsigned int y, const std::string& text, const RasterFont& font, Rgb colour, uint8_t scale)
 	{
 		drawText(x, y, unicode::utf8_to_utf32(text), font, colour, scale);
 	}
 
-	void RenderTarget::drawText(size_t x, size_t y, const std::u32string& text, const RasterFont& font, Rgb colour, uint8_t scale)
+	void RenderTarget::drawText(unsigned int x, unsigned int y, const std::u32string& text, const RasterFont& font, Rgb colour, uint8_t scale)
 	{
-		size_t text_x = (x / scale);
+		unsigned int text_x = (x / scale);
 		for (const auto& c : text)
 		{
 			const auto& g = font.get(c);
-			size_t i = 0;
-			for (size_t glyph_y = 0; glyph_y != g.height; ++glyph_y)
+			unsigned int i = 0;
+			for (unsigned int glyph_y = 0; glyph_y != g.height; ++glyph_y)
 			{
-				for (size_t glyph_x = 0; glyph_x != g.width; ++glyph_x)
+				for (unsigned int glyph_x = 0; glyph_x != g.width; ++glyph_x)
 				{
 					if (g.pixels.at(i))
 					{
